backend/requirements.txt
flask==3.0.0
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
flask-caching==2.1.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
yfinance==0.2.33
gunicorn==21.2.0
